Provide a script to login via IP directly
The old login.sh is renamed to login-via-gcloud.sh, it login via the
gcloud ssh command.

The new login.sh login via the OS ssh command directly.

The difference, mainly to us live in China, is that gcloud needs a proxy
to work, so it also login through the proxy, which can be slow.  It is
faster to login directly.  However in case you need a proxy you can
still use the login-via-gcloud.sh .